Smoking might drop seeds - issue-12 (#39)
* Refactor: camelCase file names.

* Show seeds if player has them.

Also some other CSS-ifying.

* Smoking might drop seeds.

Chance of smoking dropping a seed is based on the base seed drop rate
multiplied by the strain's seed drop mod. One problem is that currently,
when you have seeds but you run out of weed, the weed's no longer
listed, so you don't see that you have seeds. And in fact, it's probably
removed entirely, so you probably actually lose the seeds.

* Refactor smoking weed.

I was doing too much in the reducer from a single dispatch. Now the
logic is all in the actions container, with several dispatches.

* Don't show smoke button when you can't smoke.

If the selectedWeed has no quantity, don't show the button. Or if there
isn't a selected tool.

* Filter out empty weeds from weed list.

If you don't have any weed or any seeds, remove it.

* Refactor parseQuantity().

Handle none, and include the UoM display.