Seeds can be planted - issue-14 (#40)
* Refactor and cleanup, tooltip on smoke button.

* Seeds can be planted.

If your selected weed has seeds and you have empty garden space, there's
a Plant Seed button. When you click it, it reduces the seeds of your
active weed and plants that strain in your first empty garden space.
However, if it's a schwag seed, instead of planting schwag, it plants a
random tier 1 strain.

This commit also includes a miscellaneous util to select a random item
from a given array.